# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Health Pass API",
    "description": "To view the available API endpoints, first login via the /users/login API and then refresh this page in your browser.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/credentials-issuer",
      "description": "Sandbox"
    }
  ],
  "paths": {
...
    "/credentials": {
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Create credential",
        "description": "Create a new credential. Can optionally specify an expirationDate for the credential and a list of data fields that should be obfuscated",
        "operationId": "createCredential",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-hpass-issuer-id",
            "description": "Issuer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "description": "Credential type",
            "schema": {
              "type": "string",
              "enum": [
                "string",
                "encoded"
              ]
            }
          },
          {
            "in": "query",
            "name": "output",
            "description": "Output type",
            "schema": {
              "type": "string",
              "enum": [
                "qrcode"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "Credential-schema information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credential created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCredentialResponse"
                }
              },
              "image/png": {
                "schema": {
                  "$ref": "#/components/schemas/CreateQRCredentialResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/credentials/revoked": {
      "post": {
        "tags": [
          "credentials"
        ],
        "summary": "Revoke credential",
        "description": "Invalidate a credential by revoking it",
        "operationId": "revokeCredential",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-hpass-issuer-id",
            "description": "Issuer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Information for credential to be revoked",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeCredentialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/credentials/revoked/": {
      "get": {
        "tags": [
          "credentials"
        ],
        "summary": "Get all revoked credentials",
        "description": "Get all revoked credentials.",
        "operationId": "getAllRevokedCredentials",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-hpass-issuer-id",
            "description": "Issuer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "credentialID",
            "description": "Credential ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved revoke status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeStatusOptionalResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateCredentialRequest": {
        "type": "object",
        "required": [
          "schemaID",
          "data"
        ],
        "properties": {
          "schemaID": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "obfuscation": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateCredentialResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "properties": {
              "@context": {
                "type": "object",
                "properties": {
                  "cred": {
                    "type": "string"
                  }
                }
              },
              "id": {
                "type": "string"
              },
              "type": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "issuer": {
                "type": "string"
              },
              "issuanceDate": {
                "type": "string"
              },
              "credentialSchema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              },
              "credentialSubject": {
                "type": "object"
              },
              "proof": {
                "type": "object",
                "properties": {
                  "created": {
                    "type": "string"
                  },
                  "creator": {
                    "type": "string"
                  },
                  "nonce": {
                    "type": "string"
                  },
                  "signatureValue": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              },
              "expirationDate": {
                "type": "string",
                "format": "date-time"
              },
              "obfuscation": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "CreateQRCredentialResponse": {
        "type": "string",
        "format": "binary"
      },
      "RevokeCredentialRequest": {
        "type": "object",
        "required": [
          "id",
          "reason"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "RevokeStatusOptionalResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "InvalidRequestError": {
        "description": "Missing required fields in request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ConflictError": {
        "description": "Conflict error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      }
    }
  }
}